tests/steelthread/requirements.txt
dag_cbor==0.2.2
multiformats==0.1.4.post3
py_cid==0.3.0
py_multibase==1.0.3
PyNaCl==1.5.0
pytest==7.2.0
requests==2.28.1
python-jose==3.3.0